hardware: qcom: msm8998: display: fix -Wimplicit-int-float-conversion
All of the below have loss of precision due to implicit conversions to
float (or double). Accept the imprecision with explicit casts.

Fixes the following warnings from the compiler upgrade:

hardware/qcom/display/msm8998/sdm/libs/hwc2/hwc_display.cpp:548:45:
error: implicit conversion from 'uint32_t' (aka 'unsigned int') to
'float' may lose precision [-Werror,-Wimplicit-int-float-conversion]
      layer->src_rect.right = layer_buffer->width;
                            ~ ~~~~~~~~~~~~~~^~~~~
hardware/qcom/display/msm8998/sdm/libs/hwc2/hwc_display.cpp:549:46:
error: implicit conversion from 'uint32_t' (aka 'unsigned int') to
'float' may lose precision [-Werror,-Wimplicit-int-float-conversion]
      layer->src_rect.bottom = layer_buffer->height;
                             ~ ~~~~~~~~~~~~~~^~~~~~
hardware/qcom/display/msm8998/sdm/libs/hwc2/hwc_display.cpp:1718:18:
error: implicit conversion from 'uint32_t' (aka 'unsigned int') to
'float' may lose precision [-Werror,-Wimplicit-int-float-conversion]
    rect.right = primary_width;
               ~ ^~~~~~~~~~~~~
hardware/qcom/display/msm8998/sdm/libs/hwc2/hwc_display.cpp:1719:19:
error: implicit conversion from 'uint32_t' (aka 'unsigned int') to
'float' may lose precision [-Werror,-Wimplicit-int-float-conversion]
    rect.bottom = primary_height;
                ~ ^~~~~~~~~~~~~~
hardware/qcom/display/msm8998/sdm/libs/core/display_base.cpp:204:56:
error: implicit conversion from 'uint32_t' (aka 'unsigned int') to
'float' may lose precision [-Werror,-Wimplicit-int-float-conversion]
  if (gpu_target_layer_dst_xpixels > mixer_attributes_.width ||
                                   ~ ~~~~~~~~~~~~~~~~~~^~~~~
hardware/qcom/display/msm8998/sdm/libs/core/display_base.cpp:205:54:
error: implicit conversion from 'uint32_t' (aka 'unsigned int') to
'float' may lose precision [-Werror,-Wimplicit-int-float-conversion]
    gpu_target_layer_dst_ypixels > mixer_attributes_.height) {
                                 ~ ~~~~~~~~~~~~~~~~~~^~~~~~
hardware/qcom/display/msm8998/sdm/libs/core/display_base.cpp:1182:42:
error: implicit conversion from 'uint32_t' (aka 'unsigned int') to
'float' may lose precision [-Werror,-Wimplicit-int-float-conversion]
                                         layer_height), align_x);
                                         ^~~~~~~~~~~~
hardware/qcom/display/msm8998/sdm/libs/core/resource_default.cpp:472:33:
error: implicit conversion from 'uint32_t' (aka 'unsigned int') to
'float' may lose precision [-Werror,-Wimplicit-int-float-conversion]
  if ((src_width > hw_res_info_.max_pipe_width) || (dst_width > hw_res_info_.max_pipe_width)) {
                 ~ ~~~~~~~~~~~~~^~~~~~~~~~~~~~
hardware/qcom/display/msm8998/sdm/libs/core/resource_default.cpp:472:78:
error: implicit conversion from 'uint32_t' (aka 'unsigned int') to
'float' may lose precision [-Werror,-Wimplicit-int-float-conversion]
  if ((src_width > hw_res_info_.max_pipe_width) || (dst_width > hw_res_info_.max_pipe_width)) {
                                                              ~ ~~~~~~~~~~~~~^~~~~~~~~~~~~~
hardware/qcom/display/msm8998/sdm/libs/core/fb/hw_device.cpp:1336:57:
error: implicit conversion from 'uint32_t' (aka 'unsigned int') to
'float' may lose precision [-Werror,-Wimplicit-int-float-conversion]
  float max_scale_up = hw_resource_.hw_dest_scalar_info.max_scale_up;
        ~~~~~~~~~~~~   ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~~~~~~~~~
hardware/qcom/display/msm8998/sdm/libs/core/drm/hw_device_drm.cpp:763:57:
error: implicit conversion from 'uint32_t' (aka 'unsigned int') to
'float' may lose precision [-Werror,-Wimplicit-int-float-conversion]
  float max_scale_up = hw_resource_.hw_dest_scalar_info.max_scale_up;
        ~~~~~~~~~~~~   ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~~~~~~~~~
hardware/qcom/display/msm8998/sdm/libs/core/display_primary.cpp:97:38:
error: implicit conversion from 'uint32_t' (aka 'unsigned int') to
'float' may lose precision [-Werror,-Wimplicit-int-float-conversion]
    src_domain.right = variable_info.x_pixels;
                     ~ ~~~~~~~~~~~~~~^~~~~~~~
hardware/qcom/display/msm8998/sdm/libs/core/display_primary.cpp:98:39:
error: implicit conversion from 'uint32_t' (aka 'unsigned int') to
'float' may lose precision [-Werror,-Wimplicit-int-float-conversion]
    src_domain.bottom = variable_info.y_pixels;
                      ~ ~~~~~~~~~~~~~~^~~~~~~~
hardware/qcom/display/msm8998/sdm/libs/hwc2/hwc_display_external.cpp:157:37:
error: implicit conversion from 'uint32_t' (aka 'unsigned int') to
'float' may lose precision [-Werror,-Wimplicit-int-float-conversion]
  uint32_t new_mixer_width = UINT32(mixer_width * FLOAT(1.0f - width_ratio));
                                    ^~~~~~~~~~~ ~
hardware/qcom/display/msm8998/sdm/libs/hwc2/hwc_display_external.cpp:158:38:
error: implicit conversion from 'uint32_t' (aka 'unsigned int') to
'float' may lose precision [-Werror,-Wimplicit-int-float-conversion]
  uint32_t new_mixer_height = UINT32(mixer_height * FLOAT(1.0f - height_ratio));
                                     ^~~~~~~~~~~~ ~
hardware/qcom/display/msm8998/sdm/libs/hwc2/hwc_layers.cpp:682:37:
error: implicit conversion from 'uint32_t' (aka 'unsigned int') to
'float' may lose precision [-Werror,-Wimplicit-int-float-conversion]
    frame_rate = RoundToStandardFPS(fps);
                 ~~~~~~~~~~~~~~~~~~ ^~~
hardware/qcom/display/msm8998/sdm/libs/core/strategy.cpp:177:47: error:
implicit conversion from 'uint32_t' (aka 'unsigned int') to 'float' may
lose precision [-Werror,-Wimplicit-int-float-conversion]
  float layer_mixer_width = mixer_attributes_.width;
        ~~~~~~~~~~~~~~~~~   ~~~~~~~~~~~~~~~~~~^~~~~
hardware/qcom/display/msm8998/sdm/libs/core/strategy.cpp:178:48: error:
implicit conversion from 'uint32_t' (aka 'unsigned int') to 'float' may
lose precision [-Werror,-Wimplicit-int-float-conversion]
  float layer_mixer_height = mixer_attributes_.height;
        ~~~~~~~~~~~~~~~~~~   ~~~~~~~~~~~~~~~~~~^~~~~~

Bug: 139945549
Test: mm

+++ msm8998/sdm/libs/core/resource_default.cpp
@@ -469,7 +469,7 @@ DisplayError ResourceDefault::SrcSplitConfig(DisplayResourceContext *display_res
   float dst_width = dst_rect.right - dst_rect.left;
 
   // Layer cannot qualify for SrcSplit if source or destination width exceeds max pipe width.
-  if ((src_width > hw_res_info_.max_pipe_width) || (dst_width > hw_res_info_.max_pipe_width)) {
+  if ((src_width > FLOAT(hw_res_info_.max_pipe_width)) || (dst_width > FLOAT(hw_res_info_.max_pipe_width))) {
     SplitRect(src_rect, dst_rect, &left_pipe->src_roi, &left_pipe->dst_roi, &right_pipe->src_roi,
               &right_pipe->dst_roi);
     left_pipe->valid = true;
